+++ docs/source/conf.py
@@ -18,11 +18,11 @@ sys.path.insert(0, os.path.abspath('../../'))
 # -- Project information -----------------------------------------------------
 
 project = 'NVDLib'
-copyright = '2022'
+copyright = '2024'
 author = 'Vehemont'
 
 # The full version, including alpha/beta/rc tags
-release = '0.7.0'
+release = '0.7.7'
 
 
 # -- General configuration ---------------------------------------------------