+++ HISTORY.md
@@ -1,6 +1,16 @@
 Release History
 ===============
 
+0.7.7 (2024-06-02)
+-------------------
+**Bugfixes**
+- Fix cwe attr in CVE by @jobselko in https://github.com/vehemont/nvdlib/pull/39
+
+**Enhancements**
+- Added CPE Match String API with `nvdlib.searchCPEmatch`, check out the documentation at https://nvdlib.com/en/latest/v2/CVEv2.html#CPE-Match-Criteria-API and more information can be found at the NVD API documentation page: https://nvd.nist.gov/developers/products
+
+Thanks to @jobselko for submitting a PR to resolve the cwe bug.
+
 0.7.6 (2023-11-09)
 -------------------
 **Bugfixes**